requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.3
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
passlib==1.7.1
psycopg2-binary==2.7.5
pycparser==2.18
PyJWT==1.6.4
six==1.11.0
typing==3.5.3.0
Werkzeug==0.14.1
